server/requirements.txt
fastapi==0.86.0
uvicorn==0.19.0
python-dotenv==0.21.0
motor==3.1.1

